Fourth commit for Forge 1.20.1
It's 3am, I'm going to bed. I think the renderer+features are fine now. iDogEntity is mostly done at this point, I just need to work on the packets.

So, TODO: Packets/Screen
And then the small things, like the iDogBox, Calm4 Disc, and loot pool modifiers.

And then I can check and see if any of this even worked. Fingers crossed.